Delete old deserializer and supporting code
